ref: move 'setTrackMuted' to functions
