+++ react/features/base/tracks/functions.js
@@ -170,3 +170,30 @@ export function isLocalTrackMuted(tracks, mediaType) {
 
     return !track || track.muted;
 }
+
+/**
+ * Mutes or unmutes a specific <tt>JitsiLocalTrack</tt>. If the muted state of
+ * the specified <tt>track</tt> is already in accord with the specified
+ * <tt>muted</tt> value, then does nothing.
+ *
+ * @param {JitsiLocalTrack} track - The <tt>JitsiLocalTrack</tt> to mute or
+ * unmute.
+ * @param {boolean} muted - If the specified <tt>track</tt> is to be muted, then
+ * <tt>true</tt>; otherwise, <tt>false</tt>.
+ * @returns {Promise}
+ */
+export function setTrackMuted(track, muted) {
+    muted = Boolean(muted); // eslint-disable-line no-param-reassign
+
+    if (track.isMuted() === muted) {
+        return Promise.resolve();
+    }
+
+    const f = muted ? 'mute' : 'unmute';
+
+    return track[f]().catch(error => {
+
+        // FIXME emit mute failed, so that the app can show error dialog
+        console.error(`set track ${f} failed`, error);
+    });
+}
